Introduce changes necessary to bring the remainder of the manual into xml.
1. Add a <relative-path href="whatever"/> element to the manualpage
DTD.  This should give the path from the location of the page to
the root of the manual directory.  It is necessary for proper linking
to directives, etc.

2. Add the stuff to common.xsl necessary to transform a <manualpage>.
This is basically the minimum.  It creates something VERY similar to
what we had before (with the exception of a little css here and
there).  I very much hope that this will be improved soon to look
nicer.

3. Add configuring.xml, which was written by Astrid Kessler.  I made only
a very few small changes to bring it in line with the current DTD.

I have not yet committed the corresponding configuring.html.en, since I
want to give people a little opporunity to comment on this stuff.

My appologies, but I am committing a few style changes to common.xsl
along with the substantial changes.  That file is really hard to
work with, and should probably be broken up.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@96158 13f79535-47bb-0310-9956-ffa450edef68