This commit gets Perchild working again.  There are holes in this code big enough to drive a truck through, and it is NOT production quality, but I have successfully passed file descriptors between processes and served requests over the passed descriptors.  This code needs documenting, and vetting still.
Also, in order to make things work, I had to insert the socket and the
data read by the original process into the bottom of the filter stack so
that the new process could use that information.  Unfortunately, that isn't
possible to do cleanly, so I search for the CORE_IN filter, and I insert
a brigade that I have created into it's ctx pointer.  This is a hack, and
it means that Perchild _only_ works if you use the core filters.  Until I
find a better way to get that information to the bottom of the filter
stack, that is just how it is going to be.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@96430 13f79535-47bb-0310-9956-ffa450edef68