Simplify mod_dav's handler.
The old mechanism would jam a handler name in whenever DAV was
configured for a directory. i.e. there weren't really any tests other
than "is DAV handling this directory?" The name was just a marker
which was tested later.

The new mechanism simplies performs the test right in the handler.
This is now possible since all handlers are called (before, you *had*
to set your own name so that your handler would be called). Since the
test is actually quite straightforward, we actually gain overall: no
more fixups hook, and a couple integer comparisons (rather than a
string compare).

And hoo... the code is simpler, too.

This code has been verified with "litmus 0.8", my own little
regression script, and some hand testing of static pages and CGIs. DAV
continues to work, and it doesn't appear that the changes interfered
with other operation.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@96658 13f79535-47bb-0310-9956-ffa450edef68