Allow CGI scripts that use POST and mod_dav to exist in the same location.
Mod_Dav was always setting the r->handler field to "dav_handler", but
this means that mod_cgi won't run the script.  According to my reading
of the DAV RFC, mod_dav shouldn't do anything at all with a POST request,
because it is impossible to know if the POST was meant for DAV or for some
other resource.  We used to excuse POST from DAV processing, so I have
re-enabled that behavior.

PR:	13025


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@97016 13f79535-47bb-0310-9956-ffa450edef68