Add a generic locking provider to DAV - mod_dav_lock.  Other DAV providers can use the dav_hooks_locks_generic structure to implement locking in a non-optimal, but compliant fashion.  (Use 'dav-lock', 'generic', '0' to ap_lookup_provider to retrieve it.)
Use the DavGenericLockDB directive to configure the location of the lock
database (may be server-relative or absolute).

This allows all DAV providers to use a baseline locking API rather than roll
their own.

This code is based mainly off the mod_dav_fs locking code (which uses the
on-disk storage for locking), but removing components that aren't needed in a
non-file-backed repository (such as recording inodes).


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@98161 13f79535-47bb-0310-9956-ffa450edef68