Get a rid of the oversized cache. The new map-cache consists of a simple two-tiered apr_hash structure. cachep->maps contains entries for each map, which point to a hash with the actual values (map->entries).
Each map->entries hash lives in a subpool of cachep->pool.
The mtime is stored per map and if the map expires, we just clear
map->pool and create a fresh map->entries hash structure.

This removes a big chunk of code from mod_rewrite, improves readability and
even the memory footprint of the cache.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@100801 13f79535-47bb-0310-9956-ffa450edef68