Stop returning DECLINED from the output filter for proxy requests.
Filters can't return that type of status.

The way to decline filter handling is to pass through the input
data unchanged and hopefully remove the output filter 
permanently.

As long as the proxy flag was set in the fixup hook, we
didn't fill in our request config and the filter will be
a no-op anyway.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@399665 13f79535-47bb-0310-9956-ffa450edef68