Fix correct linking against PCRE. r730882 broke builds on MacOSX.
Now we put the PCRE libs into a new variable PCRE_LIBS
and add it to HTTPD_LDFLAGS.

That way the detected libs for PCRE could be reused in
other parts of the build system if necessary.

Side effect: I had to change the handling of HTTPD_FLAGS on
AIX and OS390 from hard setting of the variable to using
APR_ADDTO(). I hope that macro works on those platforms.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@730984 13f79535-47bb-0310-9956-ffa450edef68