setup cache-clean for suse, refactor server Add cron-job for suse to clean the cache of mod_crccache_client and explain in the readme how to use it
Refactor server to collect all incoming chunks into a buffer
that acts as a moving window on the response stream and feeds
it's data to the crc-read-block method to find block-matches
and the intermediate non-matched literals.
The old solution did not work well when potentially matching
blocks where not aligned on block boundaries. E.g. when the
non-matching parts had a length that differed from the
block-size